+++ client_src/SDLWrappers/SdlRenderer.h
@@ -25,7 +25,7 @@ class SdlRenderer {
     void renderPresent();
     void renderDrawPoint(int x, int y);
     void renderCopy(const class SdlTexture& texture,
-      const SDL_Rect* clip, int x, int y, int xscale = 1, int yscale = 1);
+      const SDL_Rect* clip, int x, int y, double xscale = 1, double yscale = 1);
 };
 
 #endif  // SDLRENDERER_H
